create loadSamples function
created loadSamples function and set the state of fishes that
we imported from sample-fishes.js file.
Created props names loadSamples and assigned this function on Inventory component.
We add it to Inventory component so that whenever clicking on the Load Sample Fishes
button, loadSamples function run and render the data on App component that is our
parent component.